requirements.txt
asgiref==3.3.4
certifi==2022.6.15
charset-normalizer==2.1.0
Django==3.2
django-cors-headers==3.13.0
django-tenants==3.3.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
idna==3.3
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.25.10